+++ app/api/CardanoClientApi.js
@@ -68,6 +68,12 @@ import { LOVELACES_PER_ADA } from '../config/numbersConfig';
 //   console.log('accounts:', JSON.stringify(accounts, null, 2));
 // })();
 
+// Get all wallets
+// (async () => {
+//   const wallets = await ClientApi.getWallets();
+//   console.log('wallets:', JSON.stringify(wallets, null, 2));
+// })();
+
 // Create account
 // (async () => {
 //   const account = await ClientApi.newAccount(
@@ -195,11 +201,10 @@ export default class CardanoClientApi {
 
   async getTransactions(request: GetTransactionsRequest) {
     Log.debug('CardanoClientApi::searchHistory called: ', JSON.stringify(request, null, 2));
-    const { accountId, /* walletId, */ searchTerm, skip, limit } = request;
-    // searchHistory endpoint requires accountId (account.caId) and not walletId
+    const { walletId, skip, limit } = request;
     try {
-      const history: ApiTransactions = await ClientApi.searchHistory(
-        accountId, searchTerm, skip, limit
+      const history: ApiTransactions = await ClientApi.getHistoryByWallet(
+        walletId, skip, limit
       );
       Log.debug('CardanoClientApi::searchHistory success: ', JSON.stringify(history, null, 2));
       return new Promise((resolve) => resolve({
@@ -222,12 +227,12 @@ export default class CardanoClientApi {
     try {
       // 1. create wallet
       const wallet: ApiWallet = await ClientApi.newWallet(
-        name, assurance, unit, mnemonic, password || ''
-      ); // empty string must be used if no password is set ^^
+        name, assurance, unit, mnemonic, password
+      );
       Log.debug('CardanoClientApi::createWallet success: ', JSON.stringify(wallet, null, 2));
 
       // 2. create account
-      await ClientApi.newAccount(wallet.cwId, name, password || '');
+      await ClientApi.newAccount(wallet.cwId, name, password);
 
       return _createWalletFromServerData(wallet);
     } catch (error) {
@@ -252,12 +257,10 @@ export default class CardanoClientApi {
     Log.debug('CardanoClientApi::createTransaction called');
     const { sender, receiver, amount, password } = request;
     // sender must be set as accountId (account.caId) and not walletId
-    const description = 'no description provided';
-    const title = 'no title provided';
     try {
-      const response: ApiTransaction = await ClientApi.newPaymentExtended(
-        sender, receiver, amount, title, description, password || ''
-      ); // empty string must be used if no password is set ^^
+      const response: ApiTransaction = await ClientApi.newPayment(
+        sender, receiver, amount, password
+      );
       Log.debug('CardanoClientApi::createTransaction success: ', JSON.stringify(response, null, 2));
       return _createTransactionFromServerData(response);
     } catch (error) {
@@ -274,8 +277,8 @@ export default class CardanoClientApi {
     const { accountId, password } = request;
     try {
       const data: ApiAddress = await ClientApi.newWAddress(
-        accountId, password || ''
-      ); // empty string must be used if no password is set
+        accountId, password
+      );
       Log.debug('CardanoClientApi::createAddress success: ', JSON.stringify(data, null, 2));
       return _createAddressFromServerData(data);
     } catch (error) {
@@ -322,12 +325,12 @@ export default class CardanoClientApi {
     try {
       // 1. restore wallet
       const wallet: ApiWallet = await ClientApi.restoreWallet(
-        walletName, 'CWANormal', 0, recoveryPhrase, walletPassword || ''
-      ); // empty string must be used if no password is set ^^
+        walletName, 'CWANormal', 0, recoveryPhrase, walletPassword
+      );
       Log.debug('CardanoClientApi::restoreWallet success');
 
       // 2. create account
-      await ClientApi.newAccount(wallet.cwId, walletName, walletPassword || '');
+      await ClientApi.newAccount(wallet.cwId, walletName, walletPassword);
 
       return _createWalletFromServerData(wallet);
     } catch (error) {
@@ -347,9 +350,9 @@ export default class CardanoClientApi {
 
   async importWalletFromKey(request: ImportKeyRequest) {
     Log.debug('CardanoClientApi::importWalletFromKey called');
+    const { filePath, walletPassword } = request;
     try {
-      const { filePath, walletPassword } = request;
-      const importedWallet: ApiWallet = await ClientApi.importWallet(filePath, walletPassword || '');
+      const importedWallet: ApiWallet = await ClientApi.importWallet(filePath, walletPassword);
       Log.debug('CardanoClientApi::importWalletFromKey success');
       return _createWalletFromServerData(importedWallet);
     } catch (error) {
@@ -363,9 +366,11 @@ export default class CardanoClientApi {
 
   async redeemAda(request: RedeemAdaRequest) {
     Log.debug('CardanoClientApi::redeemAda called');
-    const { redemptionCode, walletId } = request;
+    const { redemptionCode, walletId, walletPassword } = request;
     try {
-      const response: ApiTransaction = await ClientApi.redeemAda(redemptionCode, walletId);
+      const response: ApiTransaction = await ClientApi.redeemAda(
+        redemptionCode, walletId, walletPassword
+      );
       Log.debug('CardanoClientApi::redeemAda success');
       return _createTransactionFromServerData(response);
     } catch (error) {
@@ -376,10 +381,10 @@ export default class CardanoClientApi {
 
   async redeemPaperVendedAda(request: RedeemPaperVendedAdaRequest) {
     Log.debug('CardanoClientApi::redeemAdaPaperVend called');
-    const { shieldedRedemptionKey, mnemonics, walletId } = request;
+    const { shieldedRedemptionKey, mnemonics, walletId, walletPassword } = request;
     try {
       const response: ApiTransaction = await ClientApi.redeemAdaPaperVend(
-        shieldedRedemptionKey, mnemonics, walletId
+        shieldedRedemptionKey, mnemonics, walletId, walletPassword
       );
       Log.debug('CardanoClientApi::redeemAdaPaperVend success');
       return _createTransactionFromServerData(response);
@@ -552,6 +557,7 @@ export default class CardanoClientApi {
     }
   }
 
+  // TODO: remove this function - updateWallet is used for wallet rename
   async renameWallet(request: RenameWalletRequest) {
     Log.debug('CardanoClientApi::renameWallet called: ', JSON.stringify(request, null, 2));
     const { walletId, name } = request;
@@ -569,9 +575,7 @@ export default class CardanoClientApi {
     Log.debug('CardanoClientApi::changeWalletPassword called: ', JSON.stringify(request, null, 2));
     const { walletId, oldPassword, newPassword } = request;
     try {
-      await ClientApi.changeWalletPass(
-        walletId, oldPassword || '', newPassword || ''
-      ); // empty string must be used if no password is set
+      await ClientApi.changeWalletPass(walletId, oldPassword, newPassword);
       Log.debug('CardanoClientApi::changeWalletPassword success');
       return true;
     } catch (error) {
@@ -584,8 +588,7 @@ export default class CardanoClientApi {
     Log.debug('CardanoClientApi::setWalletPassword called: ', JSON.stringify(request, null, 2));
     const { walletId, password } = request;
     try {
-      await ClientApi.changeWalletPass(walletId, '', password || '');
-      // empty string must be used if no password is set ^^
+      await ClientApi.changeWalletPass(walletId, null, password);
       Log.debug('CardanoClientApi::setWalletPassword success');
       return true;
     } catch (error) {
@@ -593,7 +596,6 @@ export default class CardanoClientApi {
       throw new GenericApiError();
     }
   }
-
 
   async testReset() {
     Log.debug('CardanoClientApi::testReset called');
@@ -655,14 +657,13 @@ const _createAddressFromServerData = action(
 
 const _createTransactionFromServerData = action(
   'CardanoClientApi::_createTransactionFromServerData', (data: ApiTransaction) => {
-    const isOutgoing = 'CTOut'; // TODO: check how to determine ctType (data.ctType.tag === 'CTOut')
     const coins = data.ctAmount.getCCoin;
     const { ctmTitle, ctmDescription, ctmDate } = data.ctMeta;
     return new WalletTransaction({
       id: data.ctId,
-      title: ctmTitle || isOutgoing ? 'Ada sent' : 'Ada received',
-      type: isOutgoing ? 'adaExpend' : 'adaIncome',
-      amount: new BigNumber(isOutgoing ? -1 * coins : coins).dividedBy(LOVELACES_PER_ADA),
+      title: ctmTitle || data.ctIsOutgoing ? 'Ada sent' : 'Ada received',
+      type: data.ctIsOutgoing ? 'adaExpend' : 'adaIncome',
+      amount: new BigNumber(data.ctIsOutgoing ? -1 * coins : coins).dividedBy(LOVELACES_PER_ADA),
       date: new Date(ctmDate * 1000),
       description: ctmDescription || '',
       numberOfConfirmations: data.ctConfirmations,