+++ app/api/index.js
@@ -24,7 +24,6 @@ export type CreateAddressRequest = {
 };
 
 export type GetTransactionsRequest = {
-  accountId: string,
   walletId: string,
   searchTerm: string,
   skip: number,
@@ -88,6 +87,7 @@ export type UpdateWalletResponse = boolean;
 export type RedeemAdaRequest = {
   redemptionCode: string,
   walletId: string,
+  walletPassword: ?string,
 };
 
 export type RedeemAdaResponse = Wallet;
@@ -96,6 +96,7 @@ export type RedeemPaperVendedAdaRequest = {
   shieldedRedemptionKey: string,
   mnemonics: string,
   walletId: string,
+  walletPassword: ?string,
 };
 
 export type RedeemPaperVendedAdaResponse = RedeemPaperVendedAdaRequest;