+++ flow/declarations/DaedalusClientApi.js
@@ -32,6 +32,7 @@ declare module 'daedalus-client-api' {
     ctConfirmations: number,
     ctId: string,
     ctInputAddrs: Array<string>,
+    ctIsOutgoing: boolean,
     ctMeta: {
       ctmDate: Date,
       ctmDescription: ?string,
@@ -83,8 +84,21 @@ declare module 'daedalus-client-api' {
   declare function newAccount(walletId: string, walletName: string, walletPassword: ?string): Promise<ApiAccount>;
 
   // Transactions
-  declare function searchHistory(senderAccountId: string, searchTerm: string, skip: number, limit: number): Promise<ApiTransactions>;
-  declare function newPaymentExtended(senderAccountId: string, receiverAddress: string, amount: string, title: string, description: ?string, password: ?string): Promise<ApiTransaction>;
+  // * getHistory has first three parameters optional
+  //    - `getHistory(wId, null, null, 1, 10)` if you want to search with wallet
+  //    - `getHistory(null, accId, null, 1, 10)` if you want to search with account
+  //    - `getHistory(null, accId, addressId, 1, 10)` if you want to search specific address
+  //    - `getHistory(wId, null, addressId, 1, 10)` if you want to search specific address
+  //    - `getHistory` - any other combination will throw an error
+  // * `getHistoryByWallet(wId, 1, 10)` - search only by wallet
+  // * `getHistoryByAccount(acId, 1, 10)` - search only by account
+  // * `getHistoryByAddress(acId, addressId, 1, 10)` - search  by account and address within
+  declare function getHistory(walletId: ?string, accountId: ?string, addressId: ?string, skip: number, limit: number): Promise<ApiTransactions>;
+  declare function getHistoryByWallet(walletId: string, skip: number, limit: number): Promise<ApiTransactions>;
+  declare function getHistoryByAccount(accountId: string, skip: number, limit: number): Promise<ApiTransactions>;
+  // declare function getHistoryByAddress(accountId: string, addressId: string, skip: number, limit: number): Promise<ApiTransactions>;
+
+  declare function newPayment(senderAccountId: string, receiverAddress: string, amount: string, password: ?string): Promise<ApiTransaction>;
 
   // Ada Redemption
   declare function redeemAda(redemptionCode: string, accountId: string, walletPassword: ?string): Promise<ApiTransaction>;
@@ -96,7 +110,7 @@ declare module 'daedalus-client-api' {
   declare function newWallet(walletName: string, assurance: string, unit: number, walletMnemonic: string, walletPassword: ?string): Promise<ApiWallet>;
   declare function deleteWallet(walletId: string): Promise<{}>;
   declare function restoreWallet(walletName: string, assurance: string, unit: number, walletMnemonic: string, walletPassword: ?string): Promise<ApiWallet>;
-  declare function updateWallet(walletId: string, walletName: string, assurance: string, unit: number): Promise<ApiWallet>; // TODO: remove as it no longer exists
+  declare function updateWallet(walletId: string, walletName: string, assurance: string, unit: number): Promise<ApiWallet>;
   declare function importWallet(filePath: string, walletPassword: ?string): Promise<ApiWallet>;
   declare function newWAddress(accountId: string, walletPassword: ?string): Promise<ApiAddress>;
   declare function changeWalletPass(walletId: string, oldPassword: ?string, newPassword: ?string): Promise<{}>;